17-02-22 add ft_putnbr_unsigned_fd to libft

+++ libft/ft_putnbr_base_fd.c
@@ -6,7 +6,7 @@
 /*   By: hbourgeo <hbourgeo@student.19.be>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/15 20:48:19 by hbourgeo          #+#    #+#             */
-/*   Updated: 2022/02/17 17:58:38 by hbourgeo         ###   ########.fr       */
+/*   Updated: 2022/02/17 19:44:03 by hbourgeo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,9 +37,9 @@ static size_t check_base(char *base)
 	return (i);
 }
 
-int ft_putnbr_base_fd (unsigned int nbr, char *base, int fd)
+int ft_putnbr_base_fd (unsigned long long int nbr, char *base, int fd)
 {
-	unsigned int	nbr_bis;
+	unsigned long long int	nbr_bis;
 	size_t			size_base;
 	static int 		size;
